Refactored meal-food create and destroy actions
