+++ controllers/meals-controller.js
@@ -1,4 +1,6 @@
 const Meal = require('../models/meal');
+const Food = require('../models/food');
+const MealFood = require('../models/meal-food');
 
 class MealsController {
   static index(request, response, next) {
@@ -20,6 +22,38 @@ class MealsController {
       } else {
         response.status(200).json(meal);
       }
+    });
+  }
+
+  static create(request, response, next) {
+    Food.find(request.params.id)
+    .then((food) => {
+      Meal.find(request.params.meal_id)
+      .then((meal) => {
+        MealFood.create(request.params.meal_id, request.params.id)
+        .then(() => {
+          let message = `Successfully added ${food[0].name} to ${meal[0].name}`
+          response.status(201).json({ message: message });
+        });
+      });
+    });
+  }
+
+  static destroy(request, response, next) {
+    Food.find(request.params.id)
+    .then((food) => {
+      Meal.find(request.params.meal_id)
+      .then((meal) => {
+        MealFood.destroy(request.params.meal_id, request.params.id)
+        .then(() => {
+          if (meal[0] && food[0]) {
+            let message = `Successfully removed ${food[0].name} from ${meal[0].name}`
+            response.status(200).json({ message: message });
+          } else {
+            response.status(404).json();
+          }
+        });
+      });
     });
   }
 }
