+++ routes/api/v1/meals.js
@@ -4,5 +4,7 @@ const MealsController = require('../../../controllers/meals-controller');
 
 router.get('/', MealsController.index);
 router.get('/:id/foods', MealsController.show);
+router.post('/:meal_id/foods/:id', MealsController.create);
+router.delete('/:meal_id/foods/:id', MealsController.destroy);
 
 module.exports = router;
